+++ src/fastx_count_threads.nim
@@ -78,7 +78,8 @@ Options:
  
     
     if args["<inputfile>"].len() == 0:
-      stderr.writeLine("Waiting for STDIN... [Ctrl-C to quit, type with --help for info].")
+      if getEnv("SEQFU_QUIET") == "":
+        stderr.writeLine("[seqfu count] Waiting for STDIN... [Ctrl-C to quit, type with --help for info].")
       files.add("-")
     else:
       for file in args["<inputfile>"]: