+++ releases/changes.md
@@ -1,1 +1,3 @@
+* Improved `seqfu stats`: added sorting option and JSON output, improved test suite
 * **bugfix** Seqfu tabulate `-d` (detabulate) was too stringent in requiring forward and reverse reads to have the same length :facepalm:
+